+++ src/backend/base/langflow/services/monitor/service.py
@@ -113,8 +113,8 @@ class MonitorService(Service):
         return self.exec_query(query)
 
     def update_message(self, message_id: int, **kwargs):
-        query = f"UPDATE messages SET {', '.join(f'{k} = {v}' for k, v in kwargs.items())} WHERE id = {message_id}"
-
+        query = f"""UPDATE messages SET {', '.join(f"{k} = '{v}'" for k, v in kwargs.items())} WHERE index = {message_id}"""
+        
         return self.exec_query(query)
 
     def add_message(self, message: MessageModel):
